create sql validator

+++ config/ai.php
@@ -4,14 +4,14 @@ return [
   'anthropic' => [
     'api_key' => env('ANTHROPIC_API_KEY'),
     'model' => env('ANTHROPIC_MODEL', 'claude-sonnet-4-20250514'),
-    'max_tokens' => (int) env('ANTHROPIC_MAX_TOKENS', 2048),
+    'max_tokens' => env('ANTHROPIC_MAX_TOKENS', 2048),
     'api_version' => '2023-06-01',
   ],
 
   'limits' => [
-    'max_rows' => (int) env('ASKSQL_MAX_ROWS', 1000),
+    'max_rows' => env('ASKSQL_MAX_ROWS', 1000),
     'max_question_length' => 2000,
-    'queries_per_hour' => (int) env('ASKSQL_QUERIES_PER_HOUR', 30),
+    'queries_per_hour' => env('ASKSQL_QUERIES_PER_HOUR', 30),
     'statement_timeout_seconds' => 5,
   ],
